finish fire v1

+++ build.ps1
@@ -9,6 +9,13 @@ if (-not (Get-Command "SDL3.dll" -ErrorAction SilentlyContinue)) {
     return
 }
 
+if (-not (Get-Command "ml64" -ErrorAction SilentlyContinue)) {
+    Write-Error "ml64.exe not found. Is it in your PATH?"
+    Write-Output "You can open a Developer PowerShell for VS 20xx from Windows Terminal to get access to the build tools."
+    Write-Output "If you do not have the build tools installed, you can get them from the Visual Studio Installer, which can be downloaded here: https://visualstudio.microsoft.com/downloads/"
+    return
+}
+
 $gameName = "asteroids86"
 
 ml64 `
@@ -22,6 +29,8 @@ ml64 `
     /debug:full `
     /entry:main `
     "/out:.\bin\$gameName.exe"
+
+if ($LASTEXITCODE -ne 0) { return }
 
 if (Test-Path 'mllink$.lnk') {
     Remove-Item 'mllink$.lnk'
